+++ jpx/src/main/java/io/jenetics/jpx/format/CompositeFormat.java
@@ -19,8 +19,6 @@
  */
 package io.jenetics.jpx.format;
 
-import static java.util.Objects.requireNonNull;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;